klambda: refactor closure function to interface representation
