+++ klambda/eval.go
@@ -11,8 +11,6 @@ const (
 	controlFlowApply
 )
 
-type inst = func(ctx *ControlFlow, ebp, esp int)
-
 type ControlFlow struct {
 	// controlFlowReturn: result = data[0]
 	// controlFlowApply: fn, args = data[0], data[1], data[2] ...
@@ -341,24 +339,6 @@ func compileDefunMacro(cc *Compiler, exp Obj, env *Env, tail bool) inst {
 	return genKLGlobalSetInst(fName, inst)
 }
 
-func genKLGlobalRefInst(sym Obj) inst {
-	return func(ctx *ControlFlow, ebp, esp int) {
-		s := mustSymbol(sym)
-		if s.function == nil {
-			fmt.Println("NOT DEFINED")
-			panic("undefined symbol:" + s.str)
-		}
-		ctx.val = s.function
-	}
-}
-
-func genKLGlobalSetInst(fName Obj, f inst) inst {
-	return func(ctx *ControlFlow, ebp, esp int) {
-		f(ctx, ebp, esp)
-		ctx.val = PrimNS2Set(fName, ctx.val)
-	}
-}
-
 func compileLetMacro(cc *Compiler, exp Obj, env *Env, tail bool) inst {
 	// (let x y z) => ((lambda x z) y)
 	x := Car(exp)
@@ -387,250 +367,4 @@ func compileTrapErrorMacro(cc *Compiler, exp Obj, env *Env, tail bool) inst {
 func compileTypeMacro(cc *Compiler, exp Obj, env *Env, tail bool) inst {
 	// (type exp _) => exp
 	return cc.compile(Car(exp), env, tail)
-}
-
-func genConstInst(v Obj) inst {
-	return func(ctx *ControlFlow, ebp, esp int) {
-		ctx.val = v
-	}
-}
-
-func genIfInst(a, b, c inst) inst {
-	return func(ctx *ControlFlow, ebp, esp int) {
-		a(ctx, ebp, esp)
-		switch ctx.val {
-		case True:
-			b(ctx, ebp, esp)
-		case False:
-			c(ctx, ebp, esp)
-		default:
-			panic("test branch for if must be a boolean")
-		}
-	}
-}
-
-func genDoInst(op1, op2 inst, tail bool) inst {
-	return func(ctx *ControlFlow, ebp, esp int) {
-		op1(ctx, ebp, esp)
-		if tail {
-			ctx.pc = op2
-			return
-		}
-		op2(ctx, ebp, esp)
-	}
-}
-
-func genLocalRefInst(idx int) inst {
-	return func(ctx *ControlFlow, ebp, esp int) {
-		ctx.val = ctx.stack[ebp+idx]
-	}
-}
-
-func genClosureRefInst(m, n int) inst {
-	return func(ctx *ControlFlow, ebp, esp int) {
-		clo := mustClosure(ctx.stack[ebp])
-		for i := 1; i < m; i++ {
-			clo = clo.parent
-		}
-		ctx.val = clo.freeVars[n]
-	}
-}
-
-func genGlobalRefInst(sym Obj) inst {
-	return func(ctx *ControlFlow, ebp, esp int) {
-		s := mustSymbol(sym)
-		if s.cora == nil {
-			fmt.Println("NOT DEFINED")
-			panic("undefined symbol:" + s.str)
-		}
-		ctx.val = s.cora
-	}
-}
-
-func genLetInst(arg, body inst, idx int) inst {
-	// fmt.Println("gen the idx ===", idx)
-	return func(ctx *ControlFlow, ebp, esp int) {
-		arg(ctx, ebp, esp)
-		// fmt.Println("eval the idx ===", idx, ObjString(ctx.val), ebp, esp, "len(stack) == ", len(ctx.stack))
-
-		ctx.stack[1+ebp+idx] = ctx.val
-		body(ctx, ebp, esp)
-	}
-}
-
-func genMakeClosureInst(op inst, nargs int, mark []posRef, nlocal int) inst {
-	return func(ctx *ControlFlow, ebp, esp int) {
-		var parent *scmClosure
-		if esp > ebp {
-			if *(ctx.stack[ebp]) == scmHeadClosure {
-				parent = mustClosure(ctx.stack[ebp])
-			}
-		}
-		// Save some variables into the closure
-		// Because they are free variables.
-		var freeVars map[int]Obj
-		if len(mark) > 0 {
-			freeVars = make(map[int]Obj, len(mark))
-			for _, p := range mark {
-				if p.Up != 1 {
-					panic("should never here")
-				}
-				k := p.Offset
-				freeVars[k] = ctx.stack[ebp+k+1]
-			}
-		}
-		ctx.val = MakeClosure(op, ebp, esp, nargs, parent, freeVars, nlocal)
-	}
-}
-
-func getRequired(o Obj) int {
-	switch *o {
-	case scmHeadClosure:
-		return mustClosure(o).params
-	case scmHeadCurry:
-		return mustCurry(o).params
-	case scmHeadNative:
-		n := MustNative(o)
-		return n.require - len(n.captured)
-	}
-	panic("not implemented" + ObjString(o))
-}
-
-func recoverCall(ctx *ControlFlow, saveSP int) {
-	ctx.stack = ctx.stack[:ctx.pos]
-	ctx.pos = saveSP
-}
-
-func Call(ctx *ControlFlow, f Obj, args ...Obj) Obj {
-	return ctx.Call(f, args...)
-}
-
-func (ctx *ControlFlow) Call(f Obj, args ...Obj) Obj {
-	saveSP := ctx.pos
-	ctx.pos = len(ctx.stack)
-	ctx.stack = append(ctx.stack, f)
-	for _, arg := range args {
-		ctx.stack = append(ctx.stack, arg)
-	}
-
-	myCall(ctx)
-	recoverCall(ctx, saveSP)
-	return ctx.val
-}
-
-func genCallInst(insts []inst, debugExp Obj, tail bool) inst {
-	return func(ctx *ControlFlow, ebp, esp int) {
-		saveSP := ctx.pos
-		ctx.pos = len(ctx.stack)
-		for _, inst := range insts {
-			inst(ctx, ebp, esp)
-			ctx.stack = append(ctx.stack, ctx.val)
-		}
-
-	again:
-
-		f := ctx.stack[ctx.pos]
-		required := getRequired(f)
-		provided := len(ctx.stack[ctx.pos+1:])
-
-		if required < provided {
-			// Prepare a new stack for calling...
-			saveSP1 := ctx.pos
-			ctx.pos = len(ctx.stack)
-			for i := saveSP1; i < saveSP1+1+required; i++ {
-				ctx.stack = append(ctx.stack, ctx.stack[i])
-			}
-			// Make the call
-			myCall(ctx)
-			// Recover the stack
-			ctx.stack[saveSP1] = ctx.val
-			to := saveSP1 + 1
-			from := saveSP1 + 1 + required
-			for from < saveSP1+1+provided {
-				ctx.stack[to] = ctx.stack[from]
-				from++
-				to++
-			}
-			ctx.stack = ctx.stack[:to]
-			ctx.pos = saveSP1
-
-			goto again
-		}
-
-		if required > provided {
-			upvalues := make([]Obj, provided)
-			copy(upvalues, ctx.stack[ctx.pos+1:])
-			ctx.val = MakeCurry(f, upvalues, required-provided)
-			recoverCall(ctx, saveSP)
-			return
-		}
-
-		if required == provided {
-			if tail {
-				f := ctx.stack[ctx.pos]
-				if *f == scmHeadClosure {
-					clo := mustClosure(f)
-					nargs := len(ctx.stack[ctx.pos:])
-					copy(ctx.stack[saveSP:saveSP+nargs], ctx.stack[ctx.pos:])
-					ctx.stack = ctx.stack[:saveSP+nargs]
-					if clo.nlocal > 0 {
-						for i := 0; i < clo.nlocal; i++ {
-							ctx.stack = append(ctx.stack, Nil)
-						}
-					}
-					// fmt.Println(".... in tail call ", clo.nlocal, ctx.stack)
-					ctx.pos = saveSP
-					ctx.pc = clo.code
-					return
-				}
-			}
-
-			myCall(ctx)
-			recoverCall(ctx, saveSP)
-		}
-	}
-}
-
-func run(ctx *ControlFlow, code inst) {
-	ctx.pc = code
-	for ctx.pc != nil {
-		tmp := ctx.pc
-		ctx.pc = nil
-		tmp(ctx, ctx.pos, len(ctx.stack))
-	}
-}
-
-func myCall(ctx *ControlFlow) {
-retry:
-	f := ctx.stack[ctx.pos]
-	switch *f {
-
-	case scmHeadNative:
-		fn := MustNative(f)
-		ctx.kind = controlFlowApply
-		fn.fn(ctx)
-		ctx.val = ctx.Get(0)
-
-	case scmHeadClosure:
-		clo := mustClosure(f)
-		if clo.nlocal > 0 {
-			for i := 0; i < clo.nlocal; i++ {
-				ctx.stack = append(ctx.stack, Nil)
-			}
-			// fmt.Println(".... in call ", clo.nlocal, ctx.stack)
-		}
-		run(ctx, clo.code)
-
-	case scmHeadCurry:
-		curry := mustCurry(f)
-		saved := append([]Obj{}, ctx.stack[ctx.pos+1:]...)
-		ctx.stack = ctx.stack[:ctx.pos]
-		ctx.stack = append(ctx.stack, curry.origin)
-		ctx.stack = append(ctx.stack, curry.upvalue...)
-		ctx.stack = append(ctx.stack, saved...)
-		f = ctx.stack[ctx.pos]
-		goto retry
-	default:
-		panic("invalid call operation on:" + ObjString(f))
-	}
 }
